rename action files to not have redundant names
